sip/zmq_logging_forwarder/requirements.txt
pyzmq>=16.0.2
bottle>=0.12.13
bjoern>=2.0.1


